Shuffle some code around and apply more styles
