+++ components/Municipality.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/ban-ts-comment */
 import styled from 'styled-components'
 import Graph from './Graph'
 import { H1 } from './Typography'
@@ -16,14 +17,12 @@ import {
   pledges as pledgedEmissions,
   paris as parisEmissions,
 } from '../data/stockholm'
+import { useState } from 'react'
 
 const GraphWrapper = styled.div`
   display: flex;
   gap: 1rem;
   flex-direction: column;
-  @media only screen and (${devices.tablet}) {
-    width: 600px;
-  }
 `
 
 const Btn = styled.button`
@@ -44,6 +43,7 @@ const Flex = styled.div`
   width: 100%;
   display: flex;
   justify-content: space-between;
+  padding: 0 1.25rem;
 `
 
 const Title = styled.h3`
@@ -52,8 +52,8 @@ const Title = styled.h3`
 `
 
 const Center = styled.div`
-width: 100%
-background-color: coral;
+  width: 100%
+  background-color: coral;
   display: flex;
   align-items: center;
   justify-content: center;
@@ -93,6 +93,63 @@ const HeaderSection = styled.div`
   margin-top: 20px;
 `
 
+const RangeContainer = styled.div`
+  margin-top: 4rem;
+  display: flex;
+  justify-content: space-between;
+  overflow-x: auto;
+  padding-bottom: 1rem;
+`
+
+const Range = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+`
+
+const Slider = styled.input`
+  appearance: slider-vertical;
+  writing-mode: bt-lr; // ie and edge
+  width: 3rem;
+  margin-top: 0.25rem;
+`
+
+const Percentage = styled.label`
+  font-size: 0.75rem;
+  margin-top: 6px;
+`
+
+const MandatePeriod = styled.div`
+  font-size: 0.75rem;
+`
+
+const StartYear = styled.div`
+  border-bottom: 1px solid white;
+  font-weight: 300;
+`
+const EndYear = styled.div`
+  font-weight: 300;
+`
+
+const Help = styled.p`
+  margin-top: 2rem;
+  line-height: 1.5rem;
+`
+
+const P = styled.p`
+  margin-top: 1.5rem;
+`
+
+const MANDATE_PERIODS = [
+  [2022, 2026],
+  [2026, 2030],
+  [2030, 2034],
+  [2034, 2038],
+  [2038, 2042],
+  [2042, 2046],
+  [2046, 2050],
+]
+
 type ShareTextFn = (name: string) => string
 const STEPS: { [index: number]: { text: string; shareText: ShareTextFn } } = {
   0: {
@@ -129,6 +186,14 @@ type Props = {
 const Municipality = (props: Props) => {
   const { step, municipality, onNextStep, onPreviousStep, coatOfArmsImage } = props
 
+  const [mandateChanges, setMandateChanges] = useState(
+    MANDATE_PERIODS.map((f) => ({
+      start: f[0],
+      end: f[1],
+      change: 1.0,
+    })),
+  )
+
   const stepConfig = STEPS[step]
   if (!stepConfig) {
     throw new Error('Render a sort of 500 page I guess')
@@ -162,6 +227,14 @@ const Municipality = (props: Props) => {
     share(municipality.Name)
   }
 
+  const handleYearChange = (index: number, value: number) => {
+    setMandateChanges((m) => {
+      const copy = [...m]
+      copy[index].change = value
+      return copy
+    })
+  }
+
   return (
     <>
       <Back />
@@ -187,23 +260,60 @@ const Municipality = (props: Props) => {
             historical={historicalEmissions}
             pledged={pledgedEmissions}
             paris={parisEmissions}
+            mandatePeriodChanges={mandateChanges}
           />
-          <Flex>
-            {onPreviousStep ? (
-              <Btn onClick={onPreviousStep}>
-                <ArrowLeft />
-                Förgående
-              </Btn>
-            ) : (
-              <div></div>
-            )}
-            {onNextStep && (
-              <Btn onClick={onNextStep}>
-                Nästa <ArrowRight />
-              </Btn>
-            )}
-          </Flex>
         </GraphWrapper>
+        <Flex>
+          {onPreviousStep ? (
+            <Btn onClick={onPreviousStep}>
+              <ArrowLeft />
+              Förgående
+            </Btn>
+          ) : (
+            <div></div>
+          )}
+          {onNextStep && (
+            <Btn onClick={onNextStep}>
+              Nästa <ArrowRight />
+            </Btn>
+          )}
+        </Flex>
+        {step > 3 && (
+          <>
+            <RangeContainer>
+              {mandateChanges.map((value, i) => (
+                <Range
+                  key={i}
+                  style={{
+                    display: 'flex',
+                    flexDirection: 'column',
+                    alignItems: 'center',
+                  }}>
+                  <MandatePeriod>
+                    <StartYear>{value.start}</StartYear>
+                    <EndYear>{value.end}</EndYear>
+                  </MandatePeriod>
+                  <Slider
+                    min={1}
+                    max={2}
+                    step={0.01}
+                    value={value.change}
+                    type="range"
+                    // @ts-ignore - this is for firefox :*(
+                    orient="vertical"
+                    onChange={(e) => handleYearChange(i, parseFloat(e.target.value))}
+                  />
+                  <Percentage>{100 - Math.round(100 / value.change)}%</Percentage>
+                </Range>
+              ))}
+            </RangeContainer>
+            <Help>
+              Med hjälp av reglagen så styr du hur stora utsläppsminskningar man behöver
+              göra per mandatperiod för att nå Parisavtalet.
+            </Help>
+            <P>Dela din graf på sociala medier.</P>
+          </>
+        )}
         <ScoreCard
           population={municipality.Population}
           budget={municipality.Budget.CO2Equivalent}
